# Store gateway mmap removal: proof of concept: replace `bufio.Reader` with buffered reader optimised for our use case