@charleskorn 2022-12-22
@56quarters @pracucci interested in your thoughts on this before I take it much further - do you think this is something worth pursuing?