@charleskorn 2022-12-23
> > @56quarters @pracucci interested in your thoughts on this before I take it much further - do you think this is something worth pursuing?
> 
> I'm not a huge fan of re-implementing something in the standard library because it creates a maintenance burden for us. I would like to run the `StreamBinaryReader` in production environments for a while before doing any more performance optimization so that we can identify where we're spending time in practice.

Makes sense to me, let's wait and see how it performs in prod first.