@pracucci 2022-12-23
Maybe it's a non sense idea, but I'm wondering if we could take advantage of `bufio.Reader.Peek()` in `newV2PostingOffsetTable()`.  The reason why we need `ResetAt()` is to go back to the previous entry in case the label name has changed. Instead of doing this lookback, I'm wondering if we could change the logic to do a look forward using `Peek()`: basically, when we read a label value, we look forward to see if the next entry has the same label name. If not, then we add the current label value to the offsets, otherwise we skip it.