@charleskorn 2022-12-23
> Maybe it's a non sense idea, but I'm wondering if we could take advantage of `bufio.Reader.Peek()` in `newV2PostingOffsetTable()`. The reason why we need `ResetAt()` is to go back to the previous entry in case the label name has changed. Instead of doing this lookback, I'm wondering if we could change the logic to do a look forward using `Peek()`: basically, when we read a label value, we look forward to see if the next entry has the same label name. If not, then we add the current label value to the offsets, otherwise we skip it.

Interesting idea. It's definitely worth a try, and has the added benefit of removing the special case for the last value of the last name.

One thing we'd need to keep in mind (and I'm writing this here primarily for my future self): in order for this to work, we'd need to copy the label value somewhere first (without allocating a string), as once we peek at the next label name, we'll invalidate the value that we've previously peeked at. It'd be interesting to benchmark and see if the additional CPU time spent copying is worth it for the reduction in seeking and discarding the buffer.