@pracucci 2022-12-23
>  in order for this to work, we'd need to copy the label value somewhere first (without allocating a string), as once we peek at the next label name, we'll invalidate the value that we've previously peeked at.

By "copy the label value" you mean the "current label name" so that we can compare it with the one we'll `Peek()`?