@charleskorn 2022-12-23
> By "copy the label value" you mean the "current label name" so that we can compare it with the one we'll `Peek()`?

The current label name is already in a safe location (`currentName`). We'd need to copy the current label value somewhere so we can store it in `t.postings` if it turns out it is the last value for the current name - once we peek at the next label name, we've moved past the current label value and so that may no longer be in the buffer.