@charleskorn 2022-12-23
I've thrown together [`charleskorn/mmap-less-index-header-reader-no-seeking`](https://github.com/grafana/mimir/compare/charleskorn/mmap-less-index-header-reader-no-seeking) to implement your idea @pracucci. ([This](https://github.com/grafana/mimir/compare/charleskorn/mmap-less-index-header-reader-no-seeking#diff-43fabf3c894d3d329092e3b5935d98cb58f741802e830287e58eaa145d8dfc6aR152) is the copying I was talking about.)

While it works, it makes the code much more complex and performance isn't better than the current implementation (tested on my machine, which has a fast SSD): 

```
name                                         old time/op    new time/op    delta
NewStreamBinaryReader/1Names1Values-10          124µs ± 1%     122µs ± 3%     ~     (p=0.151 n=5+5)
NewStreamBinaryReader/1Names10Values-10         126µs ± 2%     132µs ±16%     ~     (p=0.548 n=5+5)
NewStreamBinaryReader/1Names100Values-10        134µs ± 2%     134µs ± 1%     ~     (p=0.841 n=5+5)
NewStreamBinaryReader/1Names500Values-10        166µs ± 4%     167µs ± 1%     ~     (p=0.690 n=5+5)
NewStreamBinaryReader/1Names1000Values-10       199µs ± 4%     207µs ± 1%   +4.16%  (p=0.008 n=5+5)
NewStreamBinaryReader/1Names5000Values-10       508µs ± 3%     586µs ± 2%  +15.32%  (p=0.008 n=5+5)
NewStreamBinaryReader/20Names1Values-10         170µs ± 5%     169µs ± 3%     ~     (p=0.841 n=5+5)
NewStreamBinaryReader/20Names10Values-10        201µs ± 4%     184µs ± 2%   -8.29%  (p=0.008 n=5+5)
NewStreamBinaryReader/20Names100Values-10       365µs ± 6%     381µs ± 5%     ~     (p=0.095 n=5+5)
NewStreamBinaryReader/20Names500Values-10       931µs ± 4%    1078µs ± 2%  +15.79%  (p=0.008 n=5+5)
NewStreamBinaryReader/20Names1000Values-10     1.58ms ± 4%    1.88ms ± 3%  +18.85%  (p=0.008 n=5+5)
NewStreamBinaryReader/20Names5000Values-10     6.45ms ± 1%    8.03ms ± 1%  +24.51%  (p=0.008 n=5+5)
NewStreamBinaryReader/50Names1Values-10         281µs ± 3%     276µs ± 2%     ~     (p=0.222 n=5+5)
NewStreamBinaryReader/50Names10Values-10        387µs ± 5%     327µs ± 2%  -15.43%  (p=0.008 n=5+5)
NewStreamBinaryReader/50Names100Values-10       730µs ± 3%     775µs ± 4%   +6.04%  (p=0.008 n=5+5)
NewStreamBinaryReader/50Names500Values-10      2.11ms ± 4%    2.46ms ± 0%  +16.22%  (p=0.016 n=5+4)
NewStreamBinaryReader/50Names1000Values-10     3.69ms ± 2%    4.42ms ± 0%  +19.79%  (p=0.016 n=5+4)
NewStreamBinaryReader/50Names5000Values-10     15.8ms ± 0%    20.0ms ± 2%  +26.34%  (p=0.016 n=4+5)
NewStreamBinaryReader/100Names1Values-10        537µs ± 0%     544µs ± 5%     ~     (p=0.690 n=5+5)
NewStreamBinaryReader/100Names10Values-10       743µs ± 2%     644µs ± 5%  -13.34%  (p=0.008 n=5+5)
NewStreamBinaryReader/100Names100Values-10     1.40ms ± 5%    1.48ms ± 5%     ~     (p=0.095 n=5+5)
NewStreamBinaryReader/100Names500Values-10     4.05ms ± 2%    4.75ms ± 1%  +17.18%  (p=0.008 n=5+5)
NewStreamBinaryReader/100Names1000Values-10    7.09ms ± 1%    8.57ms ± 1%  +20.78%  (p=0.008 n=5+5)
NewStreamBinaryReader/100Names5000Values-10    31.9ms ± 0%    39.6ms ± 1%  +24.08%  (p=0.008 n=5+5)
NewStreamBinaryReader/200Names1Values-10       1.14ms ± 0%    1.16ms ± 4%     ~     (p=1.000 n=4+5)
NewStreamBinaryReader/200Names10Values-10      1.52ms ± 0%    1.33ms ± 6%  -12.27%  (p=0.016 n=4+5)
NewStreamBinaryReader/200Names100Values-10     2.78ms ± 3%    2.93ms ± 2%   +5.36%  (p=0.008 n=5+5)
NewStreamBinaryReader/200Names500Values-10     7.79ms ± 1%    9.49ms ± 7%  +21.76%  (p=0.008 n=5+5)
NewStreamBinaryReader/200Names1000Values-10    14.1ms ± 1%    17.1ms ± 1%  +21.52%  (p=0.008 n=5+5)
NewStreamBinaryReader/200Names5000Values-10    64.0ms ± 1%    79.2ms ± 1%  +23.74%  (p=0.008 n=5+5)
```

As @56quarters suggested, let's wait and see if we need to invest more time into optimising this. If it turns out we need to optimise the performance of loading the index-header further, I think your earlier idea (#3761) of having the compactor generate a minimal index-header we can read in one pass is the best place to spend our effort.